requirements.txt
# Django Web Framework
django==3.1.3

# Django Cors Headers - Used to enable CORS headers in API responses, and allow requests to be made to your API server from other origins.
django-cors-headers==3.9.0

# Django Rest Framework - Used to design API logic
djangorestframework==3.12.2

# MySQL Connector - Officially used as an interface to connect Django application to the MySQL server.
mysqlclient==2.0.1

